+++ src/App.jsx
@@ -4,7 +4,7 @@ import { Navigate, Route, Routes, useLocation } from "react-router-dom";
 import AccountPreferences from "./components/accountPreferences/account/AccountPreferences";
 import ProPreferences from "./components/accountPreferences/pro/ProPreferences";
 import PreferenceSettingEdit from "./components/PreferenceSetting/Preference";
-import Signup from "./core/auth/signup/Signup";
+// import Signup from "./core/auth/signup/Signup";
 import GeneralLoading from "./layout/general-loading/GeneralLoading";
 
 import Ticked from "./pages/Ticked/Ticked";
@@ -33,15 +33,13 @@ import Notifications from "./pages/Notifications/Notifications";
 import VaProfile from "./components/VA-Profile/VaProfile";
 import VaModal1 from "./components/VA-modal/VaModal1";
 import VASignup from "./components/va-signup/auth/signup/VASignup";
-// import Login from "./components/Login/Login";
+
 import ErrorMain from "./components/404";
 import Otp from "./core/auth/otp/Otp";
 import Sub from "./core/sub/Sub";
 
 import NewPassword from "./components/resetPassword and newPassword pages/NewPassword";
-// import ResetPassword from "./components/resetPassword and newPassword pages/ResetPassword";
 
-// import VaDasboard from "./components/vaDashboard/VaDasboard";
 import VADashboard from "./components/AssistantVA";
 import Login from "./components/Login/UserLogin";
 import VATasks from "./components/VAHome/VAHome";
@@ -61,8 +59,6 @@ import CheckMail from "./components/ResetPasswordPages/CheckMail/CheckMail";
 import NewPasswordPage from "./components/ResetPasswordPages/NewPasswordPage/NewPasswordPage";
 import ResetPasswordPage from "./components/ResetPasswordPages/ResetPage/ResetPasswordPage";
 
-// import EditAccountPage from "./components/accountSettingPages/account-setting-subpages/edit-account-page/EditAccountPage";
-
 //userProfile
 import SettingsProfilee from "./core/settings/profile/SettingsProfile";
 import Edit from "./core/settings/profile/Edit";
@@ -70,6 +66,7 @@ import ChangePassword from "./core/settings/profile/ChangePassword";
 import Subscription from "./pages/Subscription/Subscription";
 import Success from "./components/subscriptionPlan/ErrorPages/Success/Success";
 import Cancel from "./components/subscriptionPlan/ErrorPages/Cancel/Cancel";
+import UserSignUp from "./core/auth/signup/UserSignUp";
 
 function App() {
   const { user } = useContext(AuthContext);
@@ -100,25 +97,18 @@ function App() {
         <Route path="/policy" element={<PrivatePolicy />} />
         <Route path="/how-it-works" element={<HowItWorks />} />
         <Route path="/marketing" element={<Ticked />} />
-        {/* <Route
-          path="/signup"
-          element={!user ? <UserSignUp /> : <Navigate to="/dashboard" />}
-        /> */}
         <Route
           path="/login"
           element={!user ? <UserLogin /> : <Navigate to="/dashboard" />}
         />
         <Route
           path="/signup"
-          element={!user ? <Signup /> : <Navigate to="/dashboard" />}
+          element={!user ? <UserSignUp /> : <Navigate to="/dashboard" />}
         />
         {/* <Route
-          path="/login"
-          element={!user ? <Login /> : <Navigate to="/dashboard" />}
-        />
-        <Route path="/newpassword" element={<NewPassword />} />
-        {/* <Route path="/resetpassword" element={<ResetPassword />} /> */}
-        {/* <Route path="/resetpassword" element={<ResetPassword />} /> */}
+          path="/signup"
+          element={!user ? <Signup /> : <Navigate to="/dashboard" />}
+        /> */}
         <Route path="/otp" element={<Otp />} />
         <Route path="/faq" element={<FAQ />} />
         <Route path="/newpassword" element={<NewPassword />} />
