+++ src/core/auth/signup/UserSignUp.jsx
@@ -1,7 +1,6 @@
 import { Box, Divider, Stack } from "@mui/material";
 import React from "react";
-// import loginPic from "../../assets/loginPicture.svg";
-// import LoginForm from "./LoginForm";
+import SignupForm from "./SignupForm";
 
 const imgStyle = {
   objectFit: `cover`,
@@ -14,7 +13,7 @@ const UserSignUp = () => {
     <Stack height={`100vh`} direction={{ xs: `column-reverse`, md: `row` }}>
       <Box height={{ xs: `90%`, md: `100%` }} flex={{ md: 1 }}>
         <Stack alignItems={{ xs: `center`, lg: `start` }}>
-          <Box padding={`1.5rem 7rem`}>
+          <Box padding={{ xs: `1.5rem 0`, sm: `1.5rem 7rem` }}>
             <img
               src="https://res.cloudinary.com/kingsleysolomon/image/upload/v1668735681/hng/todoAppVirtualAssistant/Frame_34483_msotkx.svg"
               alt="LOGO"
@@ -23,7 +22,7 @@ const UserSignUp = () => {
         </Stack>
         <Divider />
         {/* sign up form goes here */}
-        {/* <LoginForm /> */}
+        <SignupForm />
       </Box>
       <Box height={{ xs: `15%`, md: `100%` }} flex={{ md: 1 }}>
         <img
